:building_construction: Change the project structure by modules
