+++ src/fire/mod.rs
@@ -1,0 +1,3 @@
+mod fire;
+
+pub use fire::generate;
